src/users/application/usecases/get-user.usecase.ts
import { injectable, inject } from 'tsyringe'
import { UserOutput } from '../dtos/user-output.dto'
import { UsersRepository } from '@/users/repositories/users.repository'

export namespace GetUsersUseCase {
  export type Input = {
    id: string
  }

  export type Output = UserOutput

  @injectable()
  export class UseCase {
    constructor(
      @inject('UsersRepository')
      private usersRepository: UsersRepository,
    ) {}

    async execute(input: Input): Promise<Output> {
      const user = this.usersRepository.findById(input.id)

      return user
    }
  }
}
